+++ components/UserDot.tsx
@@ -2,6 +2,7 @@ import { lightOrDark } from "../utils/LightOrDark";
 import React from "react";
 import { randomColor } from "../utils/randomColor";
 import styles from "./UserDot.module.scss";
+import { TooltipImproved } from "./TooltipImproved";
 
 // Memoization of names
 const names: Array<{ name: string; color: string }> = [];