+++ components/canvas/Canvas.tsx
@@ -23,6 +23,8 @@ import { io } from "socket.io-client";
 import { notifyOthers } from "../../services/notifyOthers";
 import { getAccessToken } from "../../auth/msalHelpers";
 import { LiveIndicator } from "../LiveIndicator";
+import { CategorizationPageButton } from "../CategorizationPageButton";
+import { resetCanvasZoomAndPosition } from "./utils/ResetCanvasZoomAndPosition";
 
 export default function Canvas(): JSX.Element {
   const ref = useRef();
@@ -36,6 +38,8 @@ export default function Canvas(): JSX.Element {
 
   const { accounts } = useMsal();
   const account = useAccount(accounts[0] || {});
+
+  useEffect(() => resetCanvasZoomAndPosition(), []);
 
   useEffect(() => {
     getAccessToken().then((accessToken) => {
@@ -155,6 +159,7 @@ export default function Canvas(): JSX.Element {
         backgroundColor: "black",
       }}
     >
+      <CategorizationPageButton userCanEdit={userCanEdit} />
       <LiveIndicator
         live={socketConnected}
         title={